[CHORE] #92 - move KeywordListActivity.kt -> KeywordGuideSelectFragment.kt
