+++ app/src/main/res/layout/fragment_keyword_guide_select.xml
@@ -10,10 +10,10 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.view.keyword.KeywordListActivity">
+        tools:context=".ui.view.keyword.guide.KeywordGuideSelectFragment">
 
         <androidx.appcompat.widget.Toolbar
-            android:id="@+id/tb_keyword_list_activity"
+            android:id="@+id/tb_keyword_guide_select"
             android:layout_width="match_parent"
             android:layout_height="?attr/actionBarSize"
             android:background="@color/white"
@@ -28,7 +28,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_gravity="center"
-                android:text="@string/keyword_setting"
+                android:text="@string/keyword_select"
                 android:textColor="@color/cod_gray"
                 android:textSize="21sp"
                 android:textStyle="bold" />
@@ -39,16 +39,17 @@
             android:id="@+id/sv_keyword_list"
             android:layout_width="match_parent"
             android:layout_height="0dp"
+            android:layout_marginBottom="16dp"
             app:layout_constraintBottom_toTopOf="@+id/btn_select_finish"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/tb_keyword_list_activity">
+            app:layout_constraintTop_toBottomOf="@+id/tb_keyword_guide_select">
 
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:id="@+id/cl_in_sv"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:paddingBottom="70dp">
+                android:paddingBottom="56dp">
 
                 <TextView
                     android:id="@+id/tv_what_is_important"
@@ -119,64 +120,6 @@
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/tv_work" />
-
-                <TextView
-                    android:id="@+id/tv_my_word"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="16dp"
-                    android:layout_marginTop="48dp"
-                    android:text="@string/my_word"
-                    android:textColor="@color/cod_gray"
-                    android:textSize="16sp"
-                    app:layout_constraintStart_toStartOf="@id/cg_work"
-                    app:layout_constraintTop_toBottomOf="@id/cg_work" />
-
-                <TextView
-                    android:id="@+id/tv_modify"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="48dp"
-                    android:layout_marginEnd="16dp"
-                    android:text="@string/modify"
-                    android:visibility="gone"
-                    android:textColor="@color/royal_blue"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintTop_toBottomOf="@id/cg_work" />
-
-                <TextView
-                    android:id="@+id/tv_complete"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="48dp"
-                    android:layout_marginEnd="16dp"
-                    android:text="@string/complete"
-                    android:visibility="gone"
-                    android:textColor="@color/royal_blue"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintTop_toBottomOf="@id/cg_work" />
-
-                <com.google.android.material.chip.ChipGroup
-                    android:id="@+id/cg_my_word"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="16dp"
-                    android:paddingHorizontal="16dp"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@id/tv_my_word">
-
-                    <com.google.android.material.chip.Chip
-                        android:id="@+id/chip_add"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:stateListAnimator="@null"
-                        app:chipBackgroundColor="@android:color/transparent"
-                        app:chipIcon="@drawable/ic_add"
-                        app:chipIconSize="32dp"
-                        app:rippleColor="@android:color/transparent" />
-
-                </com.google.android.material.chip.ChipGroup>
 
             </androidx.constraintlayout.widget.ConstraintLayout>
