+++ app/src/main/res/layout/fragment_keyword_select.xml
@@ -37,7 +37,7 @@
 
         <TextView
             android:id="@+id/tv_select_method"
-            android:layout_width="wrap_content"
+            android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:text="@string/msg_select_keyword_method"
             android:textColor="@color/mainBlack"
